pyproject.toml
[project]
name = "vectorizingexamplebasedcoding"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "pandas>=2.2.2",
    "sentence-transformers>=3.0.1",
    "faiss-gpu>=1.7.2",
    "numpy<2",
    "openpyxl>=3.1.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/vectorizingexamplebasedcoding"]
